feat(ah): add ability to sell configured stack size
The client has a hard coded value for items that are stackable on the
auction and will only allow you to sell up to 12 for certain item types.
If the stack size is different in the database then the client expects
it will fail to list the stack.  This allows you to sell a single item
and if the slot contains 1 it will sell 1.  If the slot contains 99,
then the item will be listed as a stack and the appropriate number of
items will be removed from your inventory.  If for some reason you would
want to sell exactly one, then the stack would need to be split in your
inventory.